更新Ver. 2.0.1

+++ AICM_Client.sh
@@ -31,9 +31,30 @@ mput ${DengYu}*
 by
 EOF
 
-rm -rf ${aicmtmpdir}/${DengYu}*
+if [[ $keeplocaldata == "no" ]]; then
+    rm -rf ${aicmtmpdir}/${DengYu}*
+fi
 }
 
+function upupups () {
+lftp -u ${DYSFTPUSER},${DYSFTPPASSWD} sftp://${AICMSERVERIP}:${SFTPPORT} <<EOF
+cd ${DYSFTPBUDIR}/
+lcd ${aicmtmpdir}
+mput ${DengYu}*
+by
+EOF
+
+lftp -u ${DYSFTPUSER2},${DYSFTPPASSWD2} sftp://${AICMSERVERIP2}:${SFTPPORT2} <<EOF
+cd ${DYSFTPBUDIR2}/
+lcd ${aicmtmpdir}
+mput ${DengYu}*
+by
+EOF
+
+if [[ $keeplocaldata == "no" ]]; then
+    rm -rf ${aicmtmpdir}/${DengYu}*
+fi
+}
 
 if [[ $backupsqlservicedengyu == "yes" ]]; then
     BACKUPSQLBYDENGYU
@@ -47,4 +68,15 @@ if [[ $backupfileservicedengyu == "yes" ]]; then
     echo "$DengYuriver[AICM] FileBakcup任务未激活" >> /usr/bin/AICM/AICM.log
 fi
 upupup
+
+function whowillup () {
+if [[ $aloneserver == "yes" ]]; then
+	upupup
+	else
+	upupups
+fi
+}
+
+whowillup
+
 echo "$DengYuriver[AICM] 所有任务执行完毕" >> /usr/bin/AICM/AICM.log
